pyproject.toml
[tool.poetry]
name = "ssafy-custom-news-backend"
version = "0.1.0"
description = ""
authors = ["jiwoochris <travelandi01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
scikit-learn = "^1.5.2"
numpy = "^2.1.2"
dj-rest-auth = "^6.0.0"
django-allauth = "^65.1.0"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
